Refactors overlay component rendering
Overlay rendering was moved to a seperate component
to keep clarity. This PR also add styles to loading
and syntax error overlays of the diagram.
