+++ composer/modules/web/scss/modules/ballerina-editor.scss
@@ -343,3 +343,94 @@
     right: 25px;
   }
 }
+
+
+.syntax-error-overlay {
+  z-index: 1000;
+  text-align: center;
+  position: absolute;
+  height: 100%;
+  width: 100%;
+  background-color: rgba(60, 60, 60, 0.0);
+  .error-list {
+      position: relative;
+      text-align: left;
+      padding: 10px;
+      width:400px;
+      margin: auto;
+      top: 20px;
+      background-color: #d8d8d8;
+      .list-heading {
+          text-align: center;
+          font-size: 16px;
+          color:#CC0000;
+      }
+  }
+}
+
+.loading-overlay-enter {
+  opacity: 0.01;
+}
+
+.loading-overlay-enter.loading-overlay-enter-active {
+  opacity: 0.5;
+  transition: opacity 300ms ease-in;
+}
+
+.loading-overlay-leave {
+  opacity: 0.5;
+}
+
+.loading-overlay-leave.loading-overlay-leave-active {
+  opacity: 0.01;
+  transition: opacity 300ms ease-in;
+}
+
+.bal-file-editor-loading-container {
+  z-index: 1000;
+  text-align: center;
+  vertical-align: center;
+  position: absolute;
+  height: 100%;
+  width: 100%;
+  background-color: rgb(60, 60, 60);
+  color: white;
+  opacity: 0.5;
+}
+
+#parse-pending-loader{
+  left: calc(50% - 132px);
+  top: 50%;
+  position:absolute;
+  width:234px;
+  height:28px;
+  font-size: 30px;
+  margin:auto;
+  color: #fff;
+  font-weight: 500;
+}
+
+#parse-pending-loader:after {
+  overflow: hidden;
+  vertical-align: bottom;
+  -webkit-animation: ellipsis steps(4,end) 900ms infinite;
+  animation: ellipsis steps(4,end) 900ms infinite;
+  content: "\2026"; /* ascii code for the ellipsis character */
+  width: 0px;
+  display: block;
+  margin-top: -40px;
+  margin-left: 93px;
+  font-size: 50px;
+}
+
+@keyframes ellipsis {
+to {
+  width: 1.25em;
+}
+}
+
+@-webkit-keyframes ellipsis {
+to {
+  width: 1.25em;
+}
+}
